Create trekking website
Hey!my name is Bilal.I'm a tour guide from the high atlas mountains ,and you are very welcome to join me on an incredible experience, From the history of the Amazigh civilization and the ancient buildings, to the discovery of the high mountains and hiking to a hot bath in the Moroccan desert as well as Moroccan cuisine in . and a lot more, join me in this rewarding journey. You'll have fun, and learn so much about morocco and it's desert along the way!